Fix CLI

+++ qft/qft_sdrp.py
@@ -130,7 +130,7 @@ def main():
         n = int(sys.argv[1])
     sdrp = (1-1/math.sqrt(2))/2
     if len(sys.argv) > 2:
-        sdrp = int(sys.argv[2])
+        sdrp = float(sys.argv[2])
 
     print(bench_qrack(n, sdrp))
